Dirty commit. A lot of changes. ApiManager.cs is now DirectApiManager.cs and is a part of ExtDirectHandler project. ReflectionConfigurator now has public properties 'Namespace', 'Id', 'PreserveMethodCase'. Tests are moved to VS tests.
