--- src/ExtDirectHandler.Tests/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="4.5.10" targetFramework="net40" />
-  <package id="NUnit" version="2.5.10.11092" />
-  <package id="RhinoMocks" version="3.6" />
-  <package id="SharpTestsEx" version="1.1.1" />
-</packages>
