+++ src/ExtDirectHandler.VsTests/Configuration/LambdaConfiguratorTest.cs
@@ -1,70 +1,71 @@
 ﻿using System;
 using ExtDirectHandler.Configuration;
-using NUnit.Framework;
 using SharpTestsEx;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ExtDirectHandler.Tests.Configuration
+namespace ExtDirectHandler.VsTests.Configuration
 {
-	[TestFixture]
+	[TestClass]
 	public class LambdaConfiguratorTest
 	{
 		private LambdaConfigurator _target;
 
-		[SetUp]
+		//[TestInitialize]
+		[TestInitialize]
 		public void SetUp()
 		{
 			_target = new LambdaConfigurator();
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_register_methods_with_no_parameters_and_no_return_value()
 		{
 			_target.Register<TestClass>("Service.noParNoRet", x => x.NoParNoRet());
 			_target.GetMethodInfo("Service", "noParNoRet").Should().Be.SameInstanceAs(typeof(TestClass).GetMethod("NoParNoRet"));
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_register_methods_with_no_parameters_and_return_value()
 		{
 			_target.Register<TestClass>("Service.noParWithRet", x => x.NoParWithRet());
 			_target.GetMethodInfo("Service", "noParWithRet").Should().Be.SameInstanceAs(typeof(TestClass).GetMethod("NoParWithRet"));
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_register_methods_with_parameters_and_no_return_value()
 		{
 			_target.Register<TestClass>("Service.withParNoRet", x => x.WithParNoRet(default(int), default(string)));
 			_target.GetMethodInfo("Service", "withParNoRet").Should().Be.SameInstanceAs(typeof(TestClass).GetMethod("WithParNoRet"));
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_register_methods_with_parameters_and_return_value()
 		{
 			_target.Register<TestClass>("Service.withParWithRet", x => x.WithParWithRet(default(int), default(string)));
 			_target.GetMethodInfo("Service", "withParWithRet").Should().Be.SameInstanceAs(typeof(TestClass).GetMethod("WithParWithRet"));
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_throw_exception_when_method_not_registered()
 		{
 			Executing.This(() => _target.GetMethodInfo("Action", "NotRegistered")).Should().Throw<ArgumentException>().And.ValueOf.Message.Should().Be.EqualTo("Method 'Action.NotRegistered' not registered");
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_throw_exception_when_name_not_in_the_expected_format()
 		{
 			Executing.This(() => _target.Register<TestClass>("method", x => x.NoParNoRet())).Should().Throw<ArgumentException>().And.ValueOf.Message.Should().Be.EqualTo("Invalid name 'method'. Name must be in the form 'Action.method'");
 			Executing.This(() => _target.Register<TestClass>("Ns.Action.method", x => x.NoParNoRet())).Should().Throw<ArgumentException>().And.ValueOf.Message.Should().Be.EqualTo("Invalid name 'Ns.Action.method'. Name must be in the form 'Action.method'");
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_throw_exception_when_registering_the_same_method_name_multiple_times()
 		{
 			_target.Register<TestClass>("Action.method", x => x.NoParNoRet());
 			Executing.This(() => _target.Register<TestClass>("Action.method", x => x.NoParNoRet())).Should().Throw<ArgumentException>().And.ValueOf.Message.Should().Be.EqualTo("Method 'Action.method' already registered");
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_register_passed_type()
 		{
 			_target.Register<TestClass>("Action.methodFromBase", x => x.ToString());
