+++ src/ExtDirectHandler.VsTests/ParameterValuesParserTest.cs
@@ -2,45 +2,45 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Web;
-using NUnit.Framework;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using SharpTestsEx;
 
-namespace ExtDirectHandler.Tests
+namespace ExtDirectHandler.VsTests
 {
-	[TestFixture]
+	[TestClass]
 	public class ParameterValuesParserTest // TODO rename
 	{
 		private ParametersParser _target;
 
-		[SetUp]
+		[TestInitialize]
 		public void SetUp()
 		{
 			_target = new ParametersParser();
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_parse_json_positional_arguments()
 		{
 			object[] actual = _target.ParseJsonArray(GetType().GetMethod("ExampleMethod").GetParameters(), new JArray(new JValue("v1"), new JValue(123), new JValue(true)), new JsonSerializer());
 			actual.Should().Have.SameSequenceAs(new object[] { "v1", 123, true });
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_throw_exception_when_json_positional_argument_count_does_not_match()
 		{
 			Executing.This(() => _target.ParseJsonArray(GetType().GetMethod("ExampleMethod").GetParameters(), new JArray(new JValue("v1"), new JValue(123)), new JsonSerializer())).Should().Throw<Exception>().And.Exception.Message.Should().Be.EqualTo("Method expect 3 parameter(s), but passed 2 parameter(s)");
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_parse_json_named_arguments()
 		{
 			object[] actual = _target.ParseJsonObject(GetType().GetMethod("ExampleMethod").GetParameters(), new JObject { { "p3", new JValue(true) }, { "p2", new JValue(123) }, { "p1", new JValue("v1") } }, new JsonSerializer());
 			actual.Should().Have.SameSequenceAs(new object[] { "v1", 123, true });
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_use_type_default_value_when_json_named_arguments_are_missing()
 		{
 			object[] actual = _target.ParseJsonObject(GetType().GetMethod("ExampleMethod").GetParameters(), new JObject(), new JsonSerializer());
@@ -49,7 +49,7 @@ namespace ExtDirectHandler.Tests
 			actual[2].Should().Be.EqualTo(false);
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_handle_json_parameters_of_jtoken_type_as_expected()
 		{
 			var jObject = new JObject { { "prop", new JValue("val") } };
@@ -60,7 +60,7 @@ namespace ExtDirectHandler.Tests
 			JToken.DeepEquals((JToken)actual[1], jArray).Should().Be.True();
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_handle_form_parameters()
 		{
 			var httpInputStream = TestUtils.StubHttpInputStream();
@@ -76,7 +76,7 @@ namespace ExtDirectHandler.Tests
 			actual.Should().Have.SameSequenceAs(new object[] { "v1", "v2", httpInputStream, httpPostedFile });
 		}
 
-		[Test]
+		[TestMethod]
 		public void Should_use_type_default_value_when_form_arguments_are_missing()
 		{
 			var formData = new Dictionary<string, object> {
