+++ src/ExtDirectHandler/Properties/AssemblyInfo.cs
@@ -8,3 +8,4 @@ using System.Runtime.InteropServices;
 [assembly: Guid("9048a546-5d36-40b5-a192-dd32f41b4a3d")]
 
 [assembly: InternalsVisibleTo("ExtDirectHandler.Tests")]
+[assembly: InternalsVisibleTo("ExtDirectHandler.VsTests")]
