+++ src/MultipleDirectApi/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="MultipleDirectApi.Global" Language="C#" %>
